requirements.txt
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg-stubs==0.1.4